+++ src/Nethermind/Nethermind.Config/IBlocksConfig.cs
@@ -43,5 +43,8 @@ public interface IBlocksConfig : IConfig
     [ConfigItem(Description = "The genesis block load timeout, in milliseconds.", DefaultValue = "40000")]
     int GenesisTimeoutMs { get; set; }
 
+    [ConfigItem(Description = "The ticker that gas rewards are denominated in for processing logs", DefaultValue = "ETH", HiddenFromDocs = true)]
+    string GasToken { get; set; }
+
     byte[] GetExtraDataBytes();
 }
