+++ src/Nethermind/Nethermind.Consensus/Processing/BlockchainProcessor.cs
@@ -90,7 +90,7 @@ public sealed class BlockchainProcessor : IBlockchainProcessor, IBlockProcessing
         _blockTree.NewBestSuggestedBlock += OnNewBestBlock;
         _blockTree.NewHeadBlock += OnNewHeadBlock;
 
-        _stats = new ProcessingStats(_logger);
+        _stats = new ProcessingStats(stateReader, _logger);
     }
 
     private void OnNewHeadBlock(object? sender, BlockEventArgs e)
@@ -420,20 +420,7 @@ public sealed class BlockchainProcessor : IBlockchainProcessor, IBlockProcessing
             Metrics.RecoveryQueueSize = _recoveryQueue.Count;
             Metrics.ProcessingQueueSize = _blockQueue.Count;
 
-            Transaction[] txs = suggestedBlock.Transactions;
-            Address beneficiary = suggestedBlock.Header.GasBeneficiary;
-            Transaction lastTx = txs.Length > 0 ? txs[^1] : null;
-            bool isMev = false;
-            if (lastTx is not null && (lastTx.SenderAddress == beneficiary || _alternateMevPayees.Contains(lastTx.SenderAddress)))
-            {
-                // Mev reward with in last tx
-                beneficiary = lastTx.To;
-                isMev = true;
-            }
-            UInt256 beforeBalance = _stateReader.GetBalance(processingBranch.Root, beneficiary);
-            UInt256 afterBalance = _stateReader.GetBalance(suggestedBlock.StateRoot, beneficiary);
-            UInt256 rewards = beforeBalance < afterBalance ? afterBalance - beforeBalance : default;
-            _stats.UpdateStats(lastProcessed, blockProcessingTimeInMicrosecs, rewards, isMev);
+            _stats.UpdateStats(lastProcessed, processingBranch.Root, blockProcessingTimeInMicrosecs);
         }
 
         bool updateHead = !options.ContainsFlag(ProcessingOptions.DoNotUpdateHead);
@@ -783,13 +770,6 @@ public sealed class BlockchainProcessor : IBlockchainProcessor, IBlockProcessing
         _blockTree.NewBestSuggestedBlock -= OnNewBestBlock;
         _blockTree.NewHeadBlock -= OnNewHeadBlock;
     }
-
-    // Help identify mev blocks when doesn't follow regular pattern
-    private static HashSet<AddressAsKey> _alternateMevPayees = new()
-    {
-        new Address("0xa83114A443dA1CecEFC50368531cACE9F37fCCcb"), // Extra data as: beaverbuild.org
-        new Address("0x9FC3da866e7DF3a1c57adE1a97c9f00a70f010c8"), // Extra data as: Titan (titanbuilder.xyz)
-    };
 
     [DebuggerDisplay("Root: {Root}, Length: {BlocksToProcess.Count}")]
     private readonly struct ProcessingBranch
