+++ src/Nethermind/Nethermind.Merge.Plugin/MergePlugin.cs
@@ -363,8 +363,7 @@ public partial class MergePlugin : IConsensusWrapperPlugin, ISynchronizationPlug
                     _api.SyncPeerPool!,
                     _api.LogManager,
                     _api.Config<IBlocksConfig>().SecondsPerSlot,
-                    _api.Config<IMergeConfig>().SimulateBlockProduction,
-                    _api.Config<IMergeConfig>().ShowExtraData),
+                    _api.Config<IMergeConfig>().SimulateBlockProduction),
                 new GetPayloadBodiesByHashV1Handler(_api.BlockTree, _api.LogManager),
                 new GetPayloadBodiesByRangeV1Handler(_api.BlockTree, _api.LogManager),
                 new GetPayloadBodiesByHashV2Handler(_api.BlockTree, _api.LogManager),
