+++ src/Nethermind/Nethermind.Optimism/OptimismPlugin.cs
@@ -257,8 +257,7 @@ public class OptimismPlugin : IConsensusPlugin, ISynchronizationPlugin, IInitial
                 _api.SyncPeerPool!,
                 _api.LogManager,
                 _api.Config<IBlocksConfig>().SecondsPerSlot,
-                _api.Config<IMergeConfig>().SimulateBlockProduction,
-                _api.Config<IMergeConfig>().ShowExtraData),
+                _api.Config<IMergeConfig>().SimulateBlockProduction),
             new GetPayloadBodiesByHashV1Handler(_api.BlockTree, _api.LogManager),
             new GetPayloadBodiesByRangeV1Handler(_api.BlockTree, _api.LogManager),
             new GetPayloadBodiesByHashV2Handler(_api.BlockTree, _api.LogManager),
