+++ src/Nethermind/Nethermind.Runner/configs/chiado.cfg
@@ -25,7 +25,8 @@
   "Blocks": {
     "SecondsPerSlot": 5,
     "BlockProductionTimeoutMs": 3000,
-    "TargetBlockGasLimit": 17000000
+    "TargetBlockGasLimit": 17000000,
+    "GasToken": "XDAI"
   },
   "Aura": {
     "TxPriorityContractAddress": "0x4100000000000000000000000000000000000000",
