+++ src/Nethermind/Nethermind.Runner/configs/chiado_archive.cfg
@@ -23,7 +23,8 @@
   "Blocks": {
     "SecondsPerSlot": 5,
     "BlockProductionTimeoutMs": 3000,
-    "TargetBlockGasLimit": 17000000
+    "TargetBlockGasLimit": 17000000,
+    "GasToken": "XDAI"
   },
   "Receipt": {
     "TxLookupLimit": 0
