+++ src/Nethermind/Nethermind.Runner/configs/gnosis.cfg
@@ -22,7 +22,8 @@
   "Blocks": {
     "SecondsPerSlot": 5,
     "BlockProductionTimeoutMs": 3000,
-    "TargetBlockGasLimit": 17000000
+    "TargetBlockGasLimit": 17000000,
+    "GasToken": "XDAI"
   },
   "Mining": {
     "MinGasPrice": "1000000000"
