+++ src/Nethermind/Nethermind.Runner/configs/gnosis_archive.cfg
@@ -16,7 +16,8 @@
   "Blocks": {
     "SecondsPerSlot": 5,
     "BlockProductionTimeoutMs": 3000,
-    "TargetBlockGasLimit": 17000000
+    "TargetBlockGasLimit": 17000000,
+    "GasToken": "XDAI"
   },
   "Receipt": {
     "TxLookupLimit": 0
